14-clarified
I hope I have the right idea, so these are supposed to be solutions so you can just copy+paste them into the tutorial instead of a guide that gives you an idea of what the solution is. If so then this is a solution that you can just copy+paste and run the query to solve the chapter.